feat(schedule.js): adds notification when trying to cancel an appoint…

+++ frontend/src/components/Schedule.js
@@ -3,12 +3,17 @@ import axios from 'axios';
 import { useNavigate } from 'react-router-dom';
 import Sidebar from './Menu';
 import AppointmentPopup from './AppointmentPopup';
+import Modal from 'react-modal';
 import '../_assets/css/agenda.css';
+
+Modal.setAppElement('#root');
 
 const Schedule = () => {
     const [appointments, setAppointments] = useState([]);
     const [error, setError] = useState('');
     const [showPopup, setShowPopup] = useState(false);
+    const [modalIsOpen, setModalIsOpen] = useState(false);
+    const [selectedAppointment, setSelectedAppointment] = useState(null);
     const navigate = useNavigate();
 
     const fetchAppointments = async () => {
@@ -31,33 +36,47 @@ const Schedule = () => {
         }
     };
 
-    const deleteAppointment = async (appointment) => {
+    const deleteAppointment = async () => {
         const token = localStorage.getItem('token');
         if (!token) {
             navigate('/login');
             return;
         }
 
-        try {
-            await axios.delete('http://127.0.0.1:8000/api/delete-appointment/', {
-                headers: {
-                    'Authorization': `Token ${token}`
-                },
-                data: {
-                    date: appointment.date,
-                    start_time: appointment.start_time
-                }
-            });
-            setAppointments(appointments.filter(a => a.id_appointment !== appointment.id_appointment));
-        } catch (error) {
-            setError('Erro ao excluir agendamento');
-            console.error(error);
+        if (selectedAppointment) {
+            try {
+                await axios.delete('http://127.0.0.1:8000/api/delete-appointment/', {
+                    headers: {
+                        'Authorization': `Token ${token}`
+                    },
+                    data: {
+                        date: selectedAppointment.date,
+                        start_time: selectedAppointment.start_time
+                    }
+                });
+                setAppointments(appointments.filter(a => a.id_appointment !== selectedAppointment.id_appointment));
+                setSelectedAppointment(null);
+            } catch (error) {
+                setError('Erro ao excluir agendamento');
+                console.error(error);
+            }
         }
+        closeModal();
     };
 
     useEffect(() => {
         fetchAppointments();
     }, [navigate]);
+
+    const openModal = (appointment) => {
+        setSelectedAppointment(appointment);
+        setModalIsOpen(true);
+    };
+
+    const closeModal = () => {
+        setModalIsOpen(false);
+        setSelectedAppointment(null);
+    };
 
     return (
         <div className="profile-container">
@@ -81,7 +100,7 @@ const Schedule = () => {
                                     <p><strong>Hora:</strong> {appointment.start_time} - {appointment.end_time}</p>
                                     <button 
                                         className="delete-button"
-                                        onClick={() => deleteAppointment(appointment)}
+                                        onClick={() => openModal(appointment)}
                                     >
                                         Desmarcar
                                     </button>
@@ -97,6 +116,17 @@ const Schedule = () => {
                     onBook={fetchAppointments}
                 />
             )}
+            <Modal
+                isOpen={modalIsOpen}
+                onRequestClose={closeModal}
+                contentLabel="Confirmação de Exclusão"
+                className="modal"
+                overlayClassName="overlay"
+            >
+                <h2>Tem certeza que deseja desmarcar a consulta?</h2>
+                <button onClick={deleteAppointment} className="confirm-button">Sim</button>
+                <button onClick={closeModal} className="cancel-button">Não</button>
+            </Modal>
         </div>
     );
 };
